feat(v3): run the full conformance suite in a real browser
Extract the adapter's stage execution into engine-neutral run.ts (cli.ts is now
a thin Node stdin/stdout wrapper — behavior byte-identical, re-proven 181/181
via the shared runner) and add a Chromium conformance test: a sync script
copies the shared vectors into gitignored test fixtures, and
test:browser:conformance executes all 181 fixtures in real Chromium with
run_adapter.py's exact assertion contract (verdict, reject layer, byte-exact
accept facts). Result: 181/181 pass in-browser, ~1.5s, zero verifier changes
needed — the WebCrypto core runs unmodified in both engines.
