+++ packages/verifier/vite.config.playwright.ts
@@ -5,7 +5,8 @@ import { playwright } from "@vitest/browser-playwright";
 export default defineConfig({
   test: {
     include: ["**/*.browser.test.ts"],
-    exclude: ["**/*.browser.integration.test.ts", "**/node_modules/**"],
+    // v3-conformance has its own config (chromium-only, needs synced fixtures).
+    exclude: ["**/*.browser.integration.test.ts", "**/v3-conformance.browser.test.ts", "**/node_modules/**"],
     testTimeout: 30_000,
     browser: {
       enabled: true,
